backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
numpy==2.1.3
pymongo==4.10.1
python-dotenv==1.0.1
